% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{zudedup_info}
\alias{zudedup_info}
\title{What this build of zudedup is}
\usage{
zudedup_info()
}
\value{
A list: \code{version}, zudedup's version; \code{zufast}, the version of
the zufast headers compiled in; \code{gear_seed}, the seed of the gear table;
\code{boundary_rule}, a description of the format (design section 6);
\code{defaults}, the default \code{min}, \code{avg} and \code{max}; and \code{hashes}, a named
logical saying which algorithms this installation can use.
}
\description{
Reports the versions, the format constants and the hash algorithms
available, for bug reports and for checking that two installations chunk
alike.
}
\details{
XXH3-128, the default hash, is fast and not cryptographic.
}
\examples{
zudedup_info()
}
